refactor: organize invoice example by domain
