+++ examples/invoice-triage/bin/daemon.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env -S node --import tsx
-import { main } from "../src/main"
-import { option, splitRuntimeArgs, startupOption } from "./options"
+import { main } from "../src/invoice-main"
+import { option, splitRuntimeArgs, startupOption } from "./invoice-runtime"
 
 const { runtime, rest } = splitRuntimeArgs(process.argv.slice(2))
