+++ examples/invoice-triage/src/invoice-cli.ts
@@ -1,9 +1,11 @@
 import { createScope, type Lite } from "@pumped-fn/lite"
 import { cac } from "cac"
 import { readFile } from "node:fs/promises"
-import { dailyReport, enqueue, intake, listAuditEvents, prepareDatabase } from "./flows"
-import { databaseStartup, intakeLines } from "./ports"
-import type { DatabaseStartupMode } from "./migrations"
+import { prepareDatabase } from "./invoice-database-operations"
+import { enqueue, intake, listAuditEvents } from "./invoice-intake"
+import { dailyReport } from "./invoice-reporting"
+import { databaseStartup, intakeLines } from "./invoice-runtime"
+import type { DatabaseStartupMode } from "./invoice-migrations"
 
 export interface CliIo {
   out(line: string): void
