+++ examples/invoice-triage/src/invoice-main.ts
@@ -2,9 +2,11 @@ import { createScope } from "@pumped-fn/lite"
 import { logging, type Logging } from "@pumped-fn/lite-extension-logging"
 import { scheduler } from "@pumped-fn/lite-extension-scheduler"
 import { codex } from "@pumped-fn/sdk-codex"
-import { dailyReportJob, ingest, prepareDatabase, sendRemindersJob, watchReviewQueue } from "./flows"
-import { databaseEngine, databaseStartup, mailer, postgresDatabase, type Mailer } from "./ports"
-import type { DatabaseStartupMode } from "./migrations"
+import { prepareDatabase } from "./invoice-database-operations"
+import { ingest } from "./invoice-intake"
+import { dailyReportJob, sendRemindersJob, watchReviewQueue } from "./invoice-reporting"
+import { databaseEngine, databaseStartup, mailer, postgresDatabase, type Mailer } from "./invoice-runtime"
+import type { DatabaseStartupMode } from "./invoice-migrations"
 
 export interface MainOptions {
   databaseUrl?: string
