+++ examples/invoice-triage/src/invoice-watcher.ts
@@ -1,9 +1,9 @@
 import { createScope, type Lite } from "@pumped-fn/lite"
 import { readdir, readFile, watch } from "node:fs/promises"
 import { join } from "node:path"
-import { intake } from "./flows"
-import { intakeLines } from "./ports"
-import type { IntakeSummary } from "./types"
+import { intake } from "./invoice-intake"
+import { intakeLines } from "./invoice-runtime"
+import type { IntakeSummary } from "./invoice-types"
 
 export interface ImportFileOptions {
   extensions?: readonly Lite.Extension[]
