+++ examples/invoice-triage/tests/support/invoice-model.ts
@@ -1,5 +1,5 @@
 import type { Model } from "@pumped-fn/sdk"
-import type { Classification, Invoice } from "../../src/types"
+import type { Classification, Invoice } from "../../src/invoice-types"
 
 export const heuristic: Model = Object.freeze({
   complete: (_ctx: Parameters<Model["complete"]>[0], request: Parameters<Model["complete"]>[1]) => {
